refactor(config): extract discovery and raw TOML modules
