+++ crates/rag-rat-core/src/config/mod.rs
@@ -1209,787 +1209,6 @@ fn anchor_root_to_main_worktree(root: &Path) -> PathBuf {
     if anchored.is_dir() { anchored } else { root.to_path_buf() }
 }
 
-/// The main worktree root, derived from the git common dir (`<main>/.git`). Returns `None` outside
-/// a standard git repo (bare repo, custom `GIT_DIR`, git unavailable) so resolution falls back to
-/// `root` — never guess.
-/// The DEFAULT config path for the checkout containing `dir` — the DISCOVERY side of the
-/// governing seam. [`Config::load`]'s seam decides which config WINS once a file is loaded; this
-/// decides where the CLI LOOKS when no explicit `--config` was given. Without it, a linked
-/// worktree with no branch-local `rag-rat.toml` (the state `init`'s refusal deliberately leaves)
-/// dies at the existence check before the seam ever runs (Codex batch 9). Resolution:
-///  * a local `rag-rat.toml` exists → use it. In a linked worktree the seam then governs from main
-///    AND emits the divergence warning — routing discovery straight to main when a branch file
-///    exists would silently skip that warning, so the file's presence keeps the load going through
-///    the seam (governance is identical either way).
-///  * linked worktree, no local file at `dir` → the nearest `rag-rat.toml` in an ANCESTOR up to the
-///    linked worktree root (a subdirectory launch still finds a branch-local config → seam +
-///    warning); failing that, the MAIN worktree's `rag-rat.toml` (whether or not it exists yet — a
-///    missing-config hint should name where the config BELONGS).
-///  * main/non-git, no local file → the nearest `rag-rat.toml` in an ANCESTOR directory (bounded at
-///    the enclosing git root so a nested checkout never adopts a parent repo's config), so a launch
-///    from a SUBDIRECTORY of a rag-rat repo still finds the repo's config; failing that, the local
-///    path (the hint names it).
-///
-/// An EXPLICIT `--config` path never routes through this — a user override is taken literally.
-pub fn discover_config_path(dir: &Path) -> PathBuf {
-    let local = dir.join("rag-rat.toml");
-    if local.exists() {
-        return local;
-    }
-    match linked_worktree_main_root(dir) {
-        // Linked worktree: a BRANCH-LOCAL rag-rat.toml (at the worktree root, or an ancestor of
-        // `dir` within it) must still be found from a SUBDIRECTORY launch — it routes the load
-        // through the governing seam, which governs from main AND emits the divergence warning
-        // (jumping straight to main would skip that warning, or wrongly go dormant when main has no
-        // config). `nearest_config_at_or_above` is bounded at the enclosing git root (the linked
-        // worktree root), so it cannot escape into main or a parent. No branch-local config
-        // ANYWHERE in the linked worktree ⇒ main's config path (even if missing) — the
-        // governing-seam invariant.
-        Some(main_top) =>
-            nearest_config_at_or_above(dir).unwrap_or_else(|| main_top.join("rag-rat.toml")),
-        None => nearest_config_at_or_above(dir).unwrap_or(local),
-    }
-}
-
-/// Walk upward from `dir` to the nearest directory (at or above `dir`) holding a `rag-rat.toml`,
-/// returning that file's path. `None` ⇒ no rag-rat repo at or above `dir`. The single upward-walk
-/// primitive: `discover_config_path`'s non-worktree arm uses it so a subdirectory launch inside a
-/// repo finds the repo's config, and the Claude-hook cwd→config resolver
-/// (`agent_hook::find_config`) loads the returned path.
-///
-/// The climb STOPS at the enclosing git repository root: a nested checkout or submodule that has no
-/// `rag-rat.toml` of its own must NOT bind to an indexed PARENT repo's config — that would target
-/// the wrong repository for searches and, worse, for memory writes. When `dir` is not inside a git
-/// repo there is no such boundary, so the walk runs to the filesystem root.
-pub fn nearest_config_at_or_above(dir: &Path) -> Option<PathBuf> {
-    // Resolve to an ABSOLUTE path first. A relative `dir` such as `.` has `parent() == Some("")`
-    // then `None`, so the ancestor walk would only ever inspect `.` and never climb the real
-    // filesystem tree (the callers pass `Path::new(".")` for cwd). `canonicalize` also collapses
-    // `..`/symlinks so the climb follows the true directory chain. If it fails (a non-existent
-    // `dir`), fall back to walking `dir` as given rather than aborting discovery.
-    let absolute = dir.canonicalize().ok();
-    let start = absolute.as_deref().unwrap_or(dir);
-    // The enclosing git repo's workdir root — the ceiling the climb must not cross (canonicalized
-    // so it compares equal to the canonicalized `cur`). `None` for a non-git `dir` ⇒ no
-    // ceiling.
-    let boundary = crate::index::discover_repo(start)
-        .ok()
-        .and_then(|repo| repo.workdir().map(Path::to_path_buf))
-        .and_then(|workdir| workdir.canonicalize().ok());
-    let mut current = Some(start);
-    while let Some(cur) = current {
-        let candidate = cur.join("rag-rat.toml");
-        if candidate.is_file() {
-            return Some(candidate);
-        }
-        // At the git repo root: its `rag-rat.toml` was just checked; never climb into a parent
-        // repo.
-        if boundary.as_deref() == Some(cur) {
-            break;
-        }
-        current = cur.parent();
-    }
-    None
-}
-
-/// The MAIN worktree top for `root` when `root` sits in a LINKED git worktree — `None` when the
-/// checkout containing `root` IS the main worktree (or the layout has no designated main:
-/// bare-repo hubs, custom `GIT_DIR`, non-git dirs). This is THE linked-ness predicate — derived
-/// from git topology (the discovered checkout's WORKDIR vs the common dir's main), never from a
-/// path-equality proxy: comparing `root` itself to main falsely classifies a SUBDIRECTORY of the
-/// main worktree as linked, and root-anchoring success is defeated by a branch-only `[index]
-/// root` (Codex batch 8, findings 1+3). Both `Config::load`'s governing seam and the CLI's
-/// `init` refusal resolve linked-ness through this one helper.
-pub fn linked_worktree_main_root(root: &Path) -> Option<PathBuf> {
-    let repo = crate::index::discover_repo(root).ok()?;
-    let main = main_worktree_root(root)?;
-    let workdir = repo.workdir()?;
-    let workdir = workdir.canonicalize().unwrap_or_else(|_| workdir.to_path_buf());
-    (main != workdir).then_some(main)
-}
-
-fn main_worktree_root(root: &Path) -> Option<PathBuf> {
-    let repo = crate::index::discover_repo(root).ok()?;
-    let common_dir = repo.common_dir().canonicalize().ok()?;
-    // Only the standard `<main>/.git` layout maps cleanly to a main worktree root.
-    if common_dir.file_name()?.to_str()? != ".git" {
-        return None;
-    }
-    let main_root = common_dir.parent()?.to_path_buf();
-    main_root.is_dir().then_some(main_root)
-}
-
-/// The database path a `rag-rat.toml` WITHOUT a `database` key resolves to (A7), pure — no logging,
-/// no filesystem writes. The default is the CONSOLIDATED GLOBAL store (`data_dir()/rag-rat.sqlite`)
-/// — one database per machine — with two exceptions that keep an upgrade safe:
-///  1. A pre-existing legacy `<main_worktree>/.rag-rat/index.sqlite` is honored, so a repo indexed
-///     before the flip never silently abandons its authored memories; once `rag-rat consolidate`
-///     imports and renames the file away, resolution falls through to the global path.
-///  2. When no data dir resolves at all (no `HOME`/XDG on this platform), fall back to the legacy
-///     per-repo path rather than failing — the pre-A7 behavior.
-///
-/// `db_base` is the directory a relative/legacy path anchors to (the main worktree top — see
-/// `Config::load`). Public so the init wizard can display where a keyless config will land without
-/// loading one.
-pub fn default_database_path(
-    db_base: &Path,
-    identity_root: &Path,
-    repo_id_override: Option<&str>,
-) -> PathBuf {
-    let (path, _) = default_database_with_disposition(db_base, identity_root, repo_id_override);
-    path
-}
-
-/// How a keyless config's default database resolved — the load path warns per variant.
-enum DefaultDatabaseDisposition {
-    /// The root has NO derivable repo identity (non-git dir, unborn `git init`, no pin): the
-    /// per-root legacy path, exactly the pre-flip posture.
-    IdentityLess,
-    /// The legacy per-repo file is in use (awaiting `rag-rat consolidate`).
-    Legacy,
-    /// The global store (or the no-data-dir legacy fallback path, which does not exist on disk).
-    Global,
-    /// The global store, with a STRAY legacy file present DESPITE the `.imported` marker — an old
-    /// binary, a backup restore, or a stray process re-created it after consolidation.
-    GlobalWithStrayLegacy,
-}
-
-/// [`default_database_path`] plus the [`DefaultDatabaseDisposition`] the resolution took.
-///
-/// IDENTITY GATE (first, before everything): the global default REQUIRES a resolvable repo
-/// identity (`repo_identity::identity_is_resolvable` — a pin, or a git repo with a born HEAD).
-/// An identity-less root stays on its per-root `.rag-rat/index.sqlite` exactly as pre-flip:
-/// in the shared global store every such root would pool under the ONE `__unassigned__`
-/// placeholder scope — two fresh non-git projects would see and overwrite each other — and an
-/// unborn repo would strand its placeholder rows the moment its first commit mints a real id.
-/// Per-root, the placeholder stays a single-repo-DB concept with its existing adoption flow
-/// (first commit → the placeholder adopts in the per-root DB → consolidate when ready).
-///
-/// The `.imported` marker is a STAY-GLOBAL LATCH: once `rag-rat consolidate` has renamed the legacy
-/// file away, a keyless repo resolves to the global store even if a legacy `index.sqlite`
-/// REAPPEARS beside the marker (an old binary, a restored backup, a stray process) — otherwise the
-/// stray would silently divert the repo off the store its memories were imported into.
-fn default_database_with_disposition(
-    db_base: &Path,
-    identity_root: &Path,
-    repo_id_override: Option<&str>,
-) -> (PathBuf, DefaultDatabaseDisposition) {
-    let legacy = db_base.join(".rag-rat/index.sqlite");
-    if !crate::repo_identity::identity_is_resolvable(identity_root, repo_id_override) {
-        return (legacy, DefaultDatabaseDisposition::IdentityLess);
-    }
-    let marker = db_base.join(".rag-rat/index.sqlite.imported");
-    let global = crate::data_dir::global_database_path();
-    if marker.exists()
-        && let Some(global) = global
-    {
-        let disposition = if legacy.exists() {
-            DefaultDatabaseDisposition::GlobalWithStrayLegacy
-        } else {
-            DefaultDatabaseDisposition::Global
-        };
-        return (global, disposition);
-    }
-    if legacy.exists() {
-        return (legacy, DefaultDatabaseDisposition::Legacy);
-    }
-    (global.unwrap_or(legacy), DefaultDatabaseDisposition::Global)
-}
-
-/// The load-time wrapper around [`default_database_path`]: same resolution, plus a one-line notice
-/// per disposition — the deprecation nudge toward `rag-rat consolidate` while the legacy file is
-/// what keeps the repo off the global store, and a stray-file warning when a legacy file reappears
-/// after consolidation (it is ignored, never silently adopted). An identity-less root is SILENT:
-/// it is the pre-flip posture, and `rag-rat consolidate` refuses identity-less repos, so a nudge
-/// would dead-end.
-fn resolve_default_database(
-    db_base: &Path,
-    identity_root: &Path,
-    repo_id_override: Option<&str>,
-) -> PathBuf {
-    let (path, disposition) =
-        default_database_with_disposition(db_base, identity_root, repo_id_override);
-    match disposition {
-        DefaultDatabaseDisposition::Legacy => tracing::warn!(
-            path = %path.display(),
-            "using the legacy per-repo index at `.rag-rat/index.sqlite`; the default database is now \
-             the consolidated global store. Run `rag-rat consolidate` to import this repo's memories \
-             and switch to it."
-        ),
-        DefaultDatabaseDisposition::GlobalWithStrayLegacy => tracing::warn!(
-            "a stray `.rag-rat/index.sqlite` exists beside the `.imported` consolidation marker; \
-             ignoring it and staying on the consolidated global store. Delete the stray file (its \
-             contents were NOT imported)."
-        ),
-        DefaultDatabaseDisposition::Global | DefaultDatabaseDisposition::IdentityLess => {},
-    }
-    path
-}
-
-/// The DEFAULT legacy per-repo path a KEYLESS config at `root` would consult
-/// (`<main_worktree_top>/.rag-rat/index.sqlite`) — what `rag-rat consolidate` compares a pinned
-/// `database` path against to pick the right remedy: a pin AT this path just needs the key
-/// removed, while a CUSTOM pin must also move its file here first (keyless resolution never looks
-/// anywhere else, so removing the key alone would strand the custom file unimported).
-pub fn default_legacy_database_path(root: &Path) -> PathBuf {
-    main_worktree_root(root).unwrap_or_else(|| root.to_path_buf()).join(".rag-rat/index.sqlite")
-}
-
-fn normalize_existing_dir(path: &Path) -> Result<PathBuf, ConfigError> {
-    let absolute =
-        if path.is_absolute() { path.to_path_buf() } else { std::env::current_dir()?.join(path) };
-    let canonical = absolute.canonicalize()?;
-    if !canonical.is_dir() {
-        return Err(ConfigError::MissingDirectory(canonical));
-    }
-    Ok(canonical)
-}
-
-#[derive(Debug, Deserialize, PartialEq)]
-struct RawConfig {
-    #[serde(default)]
-    index: RawIndex,
-    #[serde(default)]
-    llm: RawLlm,
-    /// Presence-capture for the OLD `[local_ai]` table (renamed to `[llm]` in #317). Serde would
-    /// otherwise SILENTLY DROP this now-unknown table, loading every embedding setting as a
-    /// default (re-enabling FastEmbed, dropping a configured remote/runtime) on upgrade. We
-    /// capture it so `load` can reject it loudly with a migration instruction instead of
-    /// misconfiguring silently.
-    #[serde(default)]
-    local_ai: Option<toml::Value>,
-    /// Presence-capture for the OLD top-level `[dream]` table (the dream model config moved to
-    /// `[llm.dream]` / `[llm.dream.remote]`). Serde would otherwise SILENTLY DROP it, so an
-    /// upgrade from `[dream.model] enabled = true` would load `[llm.dream] enabled = false`
-    /// and run the deterministic passes only, never the model. Captured so `load` rejects it
-    /// loudly with a migration instruction instead of silently downgrading.
-    #[serde(default)]
-    dream: Option<toml::Value>,
-    #[serde(default)]
-    watch: RawWatch,
-    #[serde(default)]
-    log: RawLog,
-    #[serde(default)]
-    version_check: RawVersionCheck,
-    #[serde(default)]
-    oracle: RawOracle,
-    #[serde(default)]
-    search: RawSearch,
-    #[serde(default)]
-    memory: RawMemory,
-    #[serde(default)]
-    target_bindings: BTreeMap<String, Vec<String>>,
-    #[serde(default, rename = "target")]
-    target: Vec<RawTarget>,
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawWatch {
-    enabled: Option<bool>,
-    debounce_ms: Option<u64>,
-    max_latency_ms: Option<u64>,
-    periodic_sweep_secs: Option<u64>,
-}
-
-impl From<RawWatch> for WatchConfig {
-    fn from(raw: RawWatch) -> Self {
-        let default = WatchConfig::default();
-        Self {
-            enabled: raw.enabled.unwrap_or(default.enabled),
-            debounce_ms: raw.debounce_ms.unwrap_or(default.debounce_ms),
-            max_latency_ms: raw.max_latency_ms.unwrap_or(default.max_latency_ms),
-            periodic_sweep_secs: raw.periodic_sweep_secs.unwrap_or(default.periodic_sweep_secs),
-        }
-    }
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawLog {
-    enabled: Option<bool>,
-    level: Option<String>,
-    filter: Option<String>,
-    dir: Option<String>,
-    format: Option<String>,
-    max_file_bytes: Option<u64>,
-    retention_days: Option<u64>,
-    max_files: Option<u64>,
-}
-
-impl TryFrom<RawLog> for LogConfig {
-    type Error = ConfigError;
-
-    fn try_from(raw: RawLog) -> Result<Self, Self::Error> {
-        let d = LogConfig::default();
-        let level = match raw.level {
-            Some(s) => LogLevel::parse(&s).ok_or(ConfigError::UnknownLogLevel(s))?,
-            None => d.level,
-        };
-        let format = match raw.format {
-            Some(s) => LogFormat::parse(&s).ok_or(ConfigError::UnknownLogFormat(s))?,
-            None => d.format,
-        };
-        Ok(Self {
-            enabled: raw.enabled.unwrap_or(d.enabled),
-            level,
-            filter: raw.filter.filter(|s| !s.trim().is_empty()),
-            // `dir` is finalized in `load()` (needs the db path); raw value passed through as-is.
-            dir: raw.dir.map(PathBuf::from).unwrap_or_default(),
-            format,
-            max_file_bytes: raw.max_file_bytes.unwrap_or(d.max_file_bytes),
-            retention_days: raw.retention_days.unwrap_or(d.retention_days),
-            max_files: raw.max_files.unwrap_or(d.max_files),
-        })
-    }
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawVersionCheck {
-    enabled: Option<bool>,
-}
-
-impl From<RawVersionCheck> for VersionCheckConfig {
-    fn from(raw: RawVersionCheck) -> Self {
-        Self { enabled: raw.enabled.unwrap_or(VersionCheckConfig::default().enabled) }
-    }
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawOracle {
-    auto_run: Option<bool>,
-    auto_run_quiet_period_secs: Option<u64>,
-    auto_run_min_interval_secs: Option<u64>,
-}
-
-impl From<RawOracle> for OracleConfig {
-    fn from(raw: RawOracle) -> Self {
-        let default = OracleConfig::default();
-        Self {
-            auto_run: raw.auto_run.unwrap_or(default.auto_run),
-            auto_run_quiet_period_secs: raw
-                .auto_run_quiet_period_secs
-                .unwrap_or(default.auto_run_quiet_period_secs),
-            auto_run_min_interval_secs: raw
-                .auto_run_min_interval_secs
-                .unwrap_or(default.auto_run_min_interval_secs),
-        }
-    }
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawSearch {
-    graded_git_rerank: Option<bool>,
-}
-
-impl From<RawSearch> for SearchConfig {
-    fn from(raw: RawSearch) -> Self {
-        Self {
-            graded_git_rerank: raw
-                .graded_git_rerank
-                .unwrap_or(SearchConfig::default().graded_git_rerank),
-        }
-    }
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawMemory {
-    surface: Option<String>,
-}
-
-impl TryFrom<RawMemory> for MemoryConfig {
-    type Error = ConfigError;
-
-    fn try_from(raw: RawMemory) -> Result<Self, Self::Error> {
-        let surface = match raw.surface {
-            Some(s) => MemorySurface::parse(&s).ok_or(ConfigError::UnknownMemorySurface(s))?,
-            None => MemorySurface::default(),
-        };
-        Ok(Self { surface })
-    }
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawIndex {
-    root: Option<String>,
-    database: Option<String>,
-    /// `[index] repo_id` — pins the repo's identity for the consolidated global store instead of
-    /// deriving it from the root-commit hash. Set it for a fork that must NOT share memories with
-    /// its upstream, or a repo with no commits yet. Parsed here; consumed by
-    /// `resolve_repo_identity` in a later workstream (no effect on path resolution).
-    repo_id: Option<String>,
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawLlm {
-    #[serde(default)]
-    embedding: RawEmbedding,
-    #[serde(default)]
-    dream: RawDreamLlm,
-}
-
-impl TryFrom<RawLlm> for LlmConfig {
-    type Error = ConfigError;
-
-    fn try_from(raw: RawLlm) -> Result<Self, Self::Error> {
-        Ok(Self {
-            embedding: EmbeddingConfig::try_from(raw.embedding)?,
-            dream: DreamLlmConfig::try_from(raw.dream)?,
-        })
-    }
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawDreamLlm {
-    enabled: Option<bool>,
-    /// `[llm.dream.remote]` — absent → [`RemoteDreamConfig::default`] (a local-Ollama connect).
-    /// Unlike embeddings there is no in-process fallback, so a missing block still yields a
-    /// serving config rather than `None`.
-    remote: Option<RawRemoteDream>,
-}
-
-impl TryFrom<RawDreamLlm> for DreamLlmConfig {
-    type Error = ConfigError;
-
-    fn try_from(raw: RawDreamLlm) -> Result<Self, Self::Error> {
-        let remote = match raw.remote {
-            Some(remote) => RemoteDreamConfig::try_from(remote)?,
-            None => RemoteDreamConfig::default(),
-        };
-        Ok(Self { enabled: raw.enabled.unwrap_or_default(), remote })
-    }
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawRemoteDream {
-    backend: Option<String>,
-    endpoint: Option<String>,
-    cookbook: Option<String>,
-    model: Option<String>,
-    gpu: Option<String>,
-    auth_env: Option<String>,
-    request_timeout_s: Option<u64>,
-}
-
-impl TryFrom<RawRemoteDream> for RemoteDreamConfig {
-    type Error = ConfigError;
-
-    fn try_from(raw: RawRemoteDream) -> Result<Self, Self::Error> {
-        let default = RemoteDreamConfig::default();
-        let trimmed = |s: Option<String>| s.map(|v| v.trim().to_string()).filter(|v| !v.is_empty());
-        // The SERVER-side chat model name — required, non-empty. For ollama the ollama model name;
-        // for vLLM the HF id the server was launched with.
-        let model = raw.model.unwrap_or_default();
-        let model = model.trim();
-        if model.is_empty() {
-            return Err(ConfigError::DreamRemoteMissingModel);
-        }
-        // Which OpenAI-compatible server serves this block. Omitted → `ollama`.
-        let backend = match trimmed(raw.backend) {
-            None => RemoteBackend::default(),
-            Some(raw_backend) => RemoteBackend::from_db_str(&raw_backend)
-                .ok_or(ConfigError::RemoteBackendUnknown(raw_backend))?,
-        };
-        // Dream requests the CHAT capability; `infinity` is embed-only. Reject it at parse time so
-        // a misrouted backend never reaches the verdict client.
-        if !backend.supports_chat() {
-            return Err(ConfigError::DreamBackendCannotServeChat(backend.as_db_str().to_string()));
-        }
-        // The MODE is INFERRED from which URL field is set (mirrors embeddings): EXACTLY ONE of
-        // `endpoint` (connect) / `cookbook` (ephemeral). Both → ambiguous; neither → no server.
-        let endpoint = trimmed(raw.endpoint);
-        let cookbook = trimmed(raw.cookbook);
-        match (endpoint.is_some(), cookbook.is_some()) {
-            (true, true) | (false, false) => {
-                return Err(ConfigError::DreamRemoteModeAmbiguous);
-            },
-            _ => {},
-        }
-        // SECRET HYGIENE: reject a `user:pass@host` endpoint and direct the user to `auth_env`
-        // instead. Checked against the URL authority only, so an `@` in a path/query is fine.
-        // `cookbook` is a recipe spec, not a URL, so it is not checked.
-        if let Some(url) = endpoint.as_deref()
-            && endpoint_authority_has_userinfo(url)
-        {
-            return Err(ConfigError::DreamRemoteEndpointHasCredentials);
-        }
-        // EPHEMERAL-only: the GPU to provision. A PRESENT-but-empty value is a config error
-        // (clearer than silently dropping a meant-to-be-set key). Set with a connect
-        // `endpoint` it is meaningless → rejected. The VALUE is provider-specific and
-        // validated at provision time.
-        let gpu = match raw.gpu {
-            Some(g) => {
-                let g = g.trim();
-                if g.is_empty() {
-                    return Err(ConfigError::RemoteGpuEmpty);
-                }
-                if endpoint.is_some() {
-                    return Err(ConfigError::DreamRemoteGpuRequiresCookbook);
-                }
-                Some(g.to_string())
-            },
-            None => None,
-        };
-        // Optional — local Ollama needs no auth; trim if present.
-        let auth_env = trimmed(raw.auth_env);
-        Ok(Self {
-            backend,
-            endpoint,
-            cookbook,
-            model: model.to_string(),
-            gpu,
-            auth_env,
-            request_timeout_s: raw.request_timeout_s.unwrap_or(default.request_timeout_s),
-        })
-    }
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawEmbedding {
-    /// `model = "<model_id>" | "none"` — the embedding model selector, the registry model_id (the
-    /// HF path, e.g. `sentence-transformers/all-MiniLM-L6-v2`; no aliases, #317).
-    model: Option<String>,
-    #[serde(default)]
-    runtime: RawEmbeddingRuntime,
-    /// `[llm.embedding.remote]` — absent → no remote offload (`remote: None`).
-    remote: Option<RawRemoteEmbedding>,
-}
-
-impl TryFrom<RawEmbedding> for EmbeddingConfig {
-    type Error = ConfigError;
-
-    fn try_from(raw: RawEmbedding) -> Result<Self, Self::Error> {
-        let backend = match raw.model.as_deref() {
-            Some(value) => value.parse()?,
-            None => EmbeddingBackend::default(),
-        };
-        let remote = raw.remote.map(RemoteEmbeddingConfig::try_from).transpose()?;
-        // #317 rework: the `model = "..."` selector names the MODEL; a `[remote]` block serves THAT
-        // model over Ollama. The two no longer have to "agree on a backend" — the old
-        // model="ollama" coupling (`RemoteEmbeddingBackendMismatch` /
-        // `RemoteEmbeddingMissingConfig`) is gone. The only coherence left: Ollama can only
-        // serve TRANSFORMER models, so a `[remote]` block on a static/hash model is a
-        // misconfiguration the dim probe couldn't usefully explain — reject it here with a
-        // clear message.
-        // A `[remote]` block REQUIRES a transformer (FastEmbed) selected model — Ollama can only
-        // serve those. Reject when the selected backend is NOT FastEmbed: that covers static/hash
-        // (`registry_backend()` is `Some(Hash | Model2Vec)`) AND `model = "none"` (embeddings
-        // disabled → `registry_backend()` is `None`), which would otherwise slip through and leave
-        // a remote block that never installs or provisions anything.
-        if remote.is_some() && !matches!(backend.registry_backend(), Some(Backend::FastEmbed)) {
-            return Err(ConfigError::RemoteEmbeddingNonTransformerModel(
-                backend.as_str().to_string(),
-            ));
-        }
-        Ok(Self { backend, runtime: raw.runtime.into(), remote })
-    }
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawRemoteEmbedding {
-    model: Option<String>,
-    backend: Option<String>,
-    endpoint: Option<String>,
-    cookbook: Option<String>,
-    query_endpoint: Option<String>,
-    auth_env: Option<String>,
-    gpu: Option<String>,
-    num_ctx: Option<u32>,
-    batch_size: Option<u32>,
-    concurrency: Option<u32>,
-    max_batch_chars: Option<usize>,
-    request_timeout_s: Option<u64>,
-}
-
-/// Whether the URL's authority embeds userinfo (`user[:pass]@host`). Parses the authority as
-/// everything after `scheme://` up to the next `/`, `?`, or `#`, and reports an `@` in it. A bare
-/// host, an `@` in the path/query, or a URL with no scheme separator all return false. Used to keep
-/// credentials out of the endpoint string before it is persisted into the index meta.
-pub fn endpoint_authority_has_userinfo(endpoint: &str) -> bool {
-    let after_scheme = endpoint.split_once("://").map_or(endpoint, |(_, rest)| rest);
-    let authority = after_scheme.split(['/', '?', '#']).next().unwrap_or(after_scheme);
-    authority.contains('@')
-}
-
-/// If the cookbook spec's FIRST token is a RELATIVE recipe PATH, return the spec with that token
-/// resolved against `config_dir` (the `rag-rat.toml` directory). Returns `None` — leave the spec
-/// unchanged — for an npm-package spec (`@scope/pkg`, a bare name) or an ALREADY-ABSOLUTE path; the
-/// recipe runner (`node`/`npx`) would otherwise resolve a relative path against the process CWD
-/// (wherever reconcile/the watcher runs), giving ENOENT from a subdir or a daemon (R6).
-///
-/// "Path-shaped" = starts with `./` or `../`, OR ends in a recipe extension (`.mjs`/`.js`/`.ts`/
-/// `.mts`). Only the first whitespace token (the path) is rewritten; provider subcommand/args after
-/// it are preserved verbatim.
-fn resolve_relative_cookbook_path(cookbook: &str, config_dir: &Path) -> Option<String> {
-    let mut tokens = cookbook.split_whitespace();
-    let first = tokens.next()?;
-    let lower = first.to_ascii_lowercase();
-    let is_path_shaped = first.starts_with("./")
-        || first.starts_with("../")
-        || lower.ends_with(".mjs")
-        || lower.ends_with(".js")
-        || lower.ends_with(".ts")
-        || lower.ends_with(".mts");
-    if !is_path_shaped || Path::new(first).is_absolute() {
-        return None; // npm spec or already absolute → leave verbatim
-    }
-    let resolved = config_dir.join(first);
-    // NATIVE separator, no slash rewrite. This string is a single `argv` entry for `node`/`npx tsx
-    // <path>` (spawned directly, not through a shell), and both accept the platform-native
-    // separator — `\` on Windows, `/` on Unix — so the join output is already correct as-is.
-    // Rewriting `\`→`/` would be wrong two ways: on Windows it corrupts a verbatim
-    // extended-length prefix (canonicalize can yield `\\?\C:\repo`, where the suffix MUST stay
-    // backslash-delimited), and on Unix it would mangle a literal backslash in a filename.
-    // `to_string_lossy` touches neither.
-    let mut out = resolved.to_string_lossy().into_owned();
-    for arg in tokens {
-        out.push(' ');
-        out.push_str(arg);
-    }
-    Some(out)
-}
-
-impl TryFrom<RawRemoteEmbedding> for RemoteEmbeddingConfig {
-    type Error = ConfigError;
-
-    fn try_from(raw: RawRemoteEmbedding) -> Result<Self, Self::Error> {
-        let default = RemoteEmbeddingConfig::default();
-        let trimmed = |s: Option<String>| s.map(|v| v.trim().to_string()).filter(|v| !v.is_empty());
-        // The SERVER-side model name — required, non-empty. For ollama the ollama model name; for
-        // infinity/vLLM the HF id the server was launched with.
-        let model = raw.model.unwrap_or_default();
-        let model = model.trim();
-        if model.is_empty() {
-            return Err(ConfigError::RemoteEmbeddingMissingModel);
-        }
-        // Which OpenAI-compatible server serves this block. Omitted → `ollama` (back-compat).
-        let backend = match trimmed(raw.backend) {
-            None => RemoteBackend::default(),
-            Some(raw_backend) => RemoteBackend::from_db_str(&raw_backend)
-                .ok_or(ConfigError::RemoteBackendUnknown(raw_backend))?,
-        };
-        // The MODE is INFERRED from which URL field is set (#318): EXACTLY ONE of `endpoint`
-        // (connect) / `cookbook` (ephemeral). Both → ambiguous; neither → no server to reach.
-        let endpoint = trimmed(raw.endpoint);
-        let cookbook = trimmed(raw.cookbook);
-        match (endpoint.is_some(), cookbook.is_some()) {
-            (true, true) | (false, false) => {
-                return Err(ConfigError::RemoteEmbeddingModeAmbiguous);
-            },
-            _ => {},
-        }
-        // SECRET HYGIENE: `endpoint`/`query_endpoint` are persisted WHOLE into the (secret-free)
-        // index meta. A URL with userinfo (`https://user:token@host`) would copy that credential
-        // into the SQLite index — reject it and direct the user to `auth_env`. Checked against the
-        // URL authority only (between `scheme://` and the next `/`), so an `@` in a path/query is
-        // fine. `cookbook` is a recipe spec, not a URL, so it is not checked.
-        // EPHEMERAL: the LOCAL query box. Ephemeral chunks embed on the provisioned box, but that
-        // box is torn down after reconcile, so QUERIES embed against `query_endpoint` (a local
-        // server running the same model) — and `query_embed_config` PRESERVES the backend when it
-        // rewrites the ephemeral config to a connect-shaped query config. `DEFAULT_QUERY_ENDPOINT`
-        // is a local OLLAMA URL, so it only fits `backend = ollama`. For a non-ollama backend the
-        // default would build a query embedder posting the wrong route (`/embeddings`) / an HF
-        // model id at local Ollama → every query embed fails → silent, permanent BM25
-        // fallback. So require an explicit `query_endpoint` there. Ignored for connect
-        // (queries hit the connect endpoint).
-        let query_endpoint = if cookbook.is_some() {
-            match trimmed(raw.query_endpoint) {
-                Some(qe) => Some(qe),
-                None if backend == RemoteBackend::Ollama =>
-                    Some(DEFAULT_QUERY_ENDPOINT.to_string()),
-                None => {
-                    return Err(ConfigError::RemoteQueryEndpointRequiredForBackend {
-                        backend: backend.as_db_str(),
-                    });
-                },
-            }
-        } else {
-            None
-        };
-        for url in [endpoint.as_deref(), query_endpoint.as_deref()].into_iter().flatten() {
-            if endpoint_authority_has_userinfo(url) {
-                return Err(ConfigError::RemoteEmbeddingEndpointHasCredentials);
-            }
-        }
-        // Optional — local Ollama needs no auth; trim if present.
-        let auth_env = trimmed(raw.auth_env);
-        // EPHEMERAL-only: the GPU to provision. A PRESENT-but-empty/whitespace value is a config
-        // error (clearer than silently dropping a meant-to-be-set key). Set together with a connect
-        // `endpoint` it is meaningless — reject it rather than ignore it. The VALUE is
-        // provider-specific (Modal GPU class / RunPod gpuTypeId); we do NOT validate it against an
-        // allow-list — the provider does so at provision time.
-        let gpu = match raw.gpu {
-            Some(g) => {
-                let g = g.trim();
-                if g.is_empty() {
-                    return Err(ConfigError::RemoteGpuEmpty);
-                }
-                if endpoint.is_some() {
-                    return Err(ConfigError::RemoteGpuRequiresCookbook);
-                }
-                Some(g.to_string())
-            },
-            None => None,
-        };
-        if matches!(raw.num_ctx, Some(0)) {
-            return Err(ConfigError::RemoteEmbeddingInvalidNumCtx);
-        }
-        let is_connect = endpoint.is_some();
-        let concurrency = raw
-            .concurrency
-            .unwrap_or_else(|| RemoteEmbeddingConfig::omitted_concurrency_default(is_connect))
-            .max(1);
-        if concurrency > MAX_REMOTE_EMBEDDING_CONCURRENCY {
-            return Err(ConfigError::RemoteEmbeddingConcurrencyTooHigh {
-                value: concurrency,
-                max: MAX_REMOTE_EMBEDDING_CONCURRENCY,
-            });
-        }
-        Ok(Self {
-            model: model.to_string(),
-            backend,
-            endpoint,
-            cookbook,
-            query_endpoint,
-            auth_env,
-            gpu,
-            num_ctx: raw.num_ctx,
-            batch_size: raw.batch_size.unwrap_or(default.batch_size),
-            concurrency,
-            max_batch_chars: raw.max_batch_chars.unwrap_or(default.max_batch_chars).max(1),
-            request_timeout_s: raw.request_timeout_s.unwrap_or(default.request_timeout_s),
-        })
-    }
-}
-
-#[derive(Debug, Default, Deserialize, PartialEq)]
-struct RawEmbeddingRuntime {
-    batch_size: Option<u32>,
-    ort_threads: Option<u32>,
-    omp_threads: Option<u32>,
-    max_embedding_chars: Option<usize>,
-}
-
-impl From<RawEmbeddingRuntime> for EmbeddingRuntimeConfig {
-    fn from(raw: RawEmbeddingRuntime) -> Self {
-        let default = EmbeddingRuntimeConfig::default();
-        Self {
-            batch_size: raw.batch_size.unwrap_or(default.batch_size),
-            ort_threads: raw.ort_threads.or(default.ort_threads),
-            omp_threads: raw.omp_threads.or(default.omp_threads),
-            max_embedding_chars: raw.max_embedding_chars.unwrap_or(default.max_embedding_chars),
-        }
-    }
-}
-
-#[derive(Debug, Deserialize, PartialEq)]
-struct RawTarget {
-    name: String,
-    language: String,
-    directories: Vec<String>,
-    kind: Option<String>,
-    include: Option<Vec<String>>,
-    exclude: Option<Vec<String>>,
-}
-
 #[derive(Debug, Error)]
 pub enum ConfigError {
     #[error("failed to read config: {0}")]
@@ -2119,6 +1338,19 @@ pub enum ConfigError {
     #[error("[memory] `surface` must be `full` or `summary` (got `{0}`)")]
     UnknownMemorySurface(String),
 }
+
+mod discovery;
+mod raw;
+
+pub use discovery::{
+    default_database_path, default_legacy_database_path, discover_config_path,
+    linked_worktree_main_root, nearest_config_at_or_above,
+};
+pub(crate) use discovery::{main_worktree_root, normalize_existing_dir, resolve_default_database};
+pub use raw::endpoint_authority_has_userinfo;
+pub(crate) use raw::{RawConfig, RawTarget, resolve_relative_cookbook_path};
+#[cfg(test)]
+pub(crate) use raw::{RawMemory, RawOracle, RawSearch, RawVersionCheck, RawWatch};
 
 #[cfg(test)]
 mod tests {
